fixed additional typos and parameter values

+++ components/cronet/android/sample/src/org/chromium/cronet_sample_apk/CronetSampleApplication.java
@@ -25,8 +25,8 @@ public class CronetSampleApplication extends Application {
                 .enableHttpCache(CronetEngine.Builder.HTTP_CACHE_IN_MEMORY, 100 * 1024)
                 .enableHttp2(true)
                 .setDisabledCipherSuites("0xc024,0xc02f,0002")
-                .setMinSslVersion(771) // net::SSL_PROTOCOL_VERSION_TLS1_2 -> hex value 0x0303
-                .setMaxSslVersion(772) // net::SSL_PROTOCOL_VERSION_TLS1_3 -> hex value 0x0304
+                .setMinSslVersion((short)771) // net::SSL_PROTOCOL_VERSION_TLS1_2 -> hex value 0x0303
+                .setMaxSslVersion((short)772) // net::SSL_PROTOCOL_VERSION_TLS1_3 -> hex value 0x0304
                 .enableQuic(true);
         mCronetEngine = myBuilder.build();
     }
@@ -65,8 +65,8 @@ public class CronetSampleApplication extends Application {
                         .enableHttpCache(CronetEngine.Builder.HTTP_CACHE_IN_MEMORY, 100 * 1024)
                         .enableHttp2(true)
                         .setDisabledCipherSuites("0xc024,0xc02f,0002")
-                        .setMinSslVersion(771) // net::SSL_PROTOCOL_VERSION_TLS1_2 -> hex value 0x0303
-                        .setMaxSslVersion(772) // net::SSL_PROTOCOL_VERSION_TLS1_3 -> hex value 0x0304
+                        .setMinSslVersion((short)771) // net::SSL_PROTOCOL_VERSION_TLS1_2 -> hex value 0x0303
+                        .setMaxSslVersion((short)772) // net::SSL_PROTOCOL_VERSION_TLS1_3 -> hex value 0x0304
                         .enableQuic(true)
                         .build();
     }
